backend/requirements.txt
Flask
flask-cors
PyJWT
Werkzeug
Flask-SQLAlchemy
